requirements.txt
numpy>=2.3.2
requests>=2.32.5
torch>=2.8.0
Pillow>=11.3.0
pure-python-adb>=0.3.0.dev0